Move transport.transaction.stress to integration suit
