+++ test/integration/scenarios/index.js
@@ -7,5 +7,8 @@ module.exports = {
 	propagation: {
 		blocks: require('./propagation/blocks'),
 		transactions: require('./propagation/transactions')
+	},
+	stress: {
+		transfer: require('./stress/0.transfer')
 	}
 };
